refactor + strip unused code
