--- KeyStash/Extensions/ButtonSidebarStyle.swift
@@ -1,5 +1,0 @@
-import SwiftUI
-
-extension Button {
-	
-}
