+++ KeyStash/Extensions/StripNewLines.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension String {
+	func stripNewLines() -> String {
+		let newlines = CharacterSet.newlines
+		return components(separatedBy: newlines).joined()
+	}
+}
