+++ KeyStash/Logic/Helpers/isDownloadLink.swift
