+++ KeyStash/Logic/Helpers/stringToClipboard.swift
